requirements.txt
# PyTorch - install with CUDA support
# Use: uv pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Core ML/Data Science (Python 3.12 compatible)
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Progress & Utilities
tqdm>=4.65.0

# Web Framework (for model server)
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Google AI (if using Gemini/Google services)
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
